API Reference

Base.dump

EmbeddingsAnalysis.conceptnet2wv

EmbeddingsAnalysis.pca_reduction

EmbeddingsAnalysis.similarity_order

EmbeddingsAnalysis.conceptnet2wv — Method.

conceptnet2wv(cptnet, language)

Converts a ConceptNet object, cptnet to a WordVectors object. The language of the word embeddings has to be specified explicitly as a Symbol or Languages.Language (Conceptnet embeddings can be multilingual).

source

EmbeddingsAnalysis.pca_reduction — Method.

pca_reduction(wv::WordVectors, rdim=7, outdim=size(wv.vectors,1); [do_pca=true])

Post-processes word embeddings wv by removing the first rdim PCA components from the word vectors and also reduces the dimensionality to outdim through a subsequent PCA transform, if do_pca=true.

Arguments

wv::WordVectors the word embeddings

rdim::Int the number of PCA components to remove from the data (default 7)

outdim::Int the output dimensionality of the data after the PCA dimensionality reduction; it is performed only if do_pca=true and the default value is the same as that of the input embeddings i.e. no reduction

Keyword arguments

do_pca::Bool whether to perform a PCA transform of the post-processed data (default true)

References:

Vikas Raunak "Simple and effective dimensionality reduction for word embeddings", NIPS 2017 Workshop

source

EmbeddingsAnalysis.similarity_order — Method.

similarity_order(wv::WordVectors, alpha=-0.65)

Post-processes the word embeddings wv so that the embeddings capture more information than directly apparent through a linear transformation that adjusts the similarity order of the model. The function returns a new WordVectors object containing the processed embeddings.

Arguments

wv::WordVectors the word embeddings

alpha::AbstractFloat the α parameter of the algorithm (default -0.65)

References:

Artetxe et al. "Uncovering divergent linguistic information in word embeddings with lessons for intrinsic and extrinsic evaluation", 2018

source

Base.dump — Method.

dump(filename::AbstractString, wv::WordVectors [;kind=:binary])

Writes embeddings to disk.

Arguments

filename::AbstractString the embeddings file name

wv::WordVectors the embeddings

Keyword arguments

kind::Symbol specifies whether the embeddings file is textual (:text)

or binary (:binary); default :binary

source