API Reference

Base.dump

EmbeddingsAnalysis.similarity_order

Base.dump — Method.

dump(wv::WordVectors, filename [;kind=:binary])

Writes embeddings to disk.

Arguments

wv::WordVectors the embeddings

filename::AbstractString the embeddings file name

Keyword arguments

kind::Symbol specifies whether the embeddings file is textual (:text)

or binary (:binary); default :binary

source

EmbeddingsAnalysis.similarity_order — Method.

similarity_order(wv::WordVectors, alpha=-0.65)

Post-processes the word embeddings wv so that the embeddings capture more information than directly apparent through a linear transformation that adjusts the similarity order of the model. The function returns a new WordVectors object containing the processed embeddings.

References:

Artetxe et al. "Uncovering divergent linguistic information in word embeddings with lessons for intrinsic and extrinsic evaluation", 2018

source